Tighten skill install internals
Keep `install_source` crate-visible, carry selected skill install
sources as a typed enum, and split the web Skills handler tests into
smaller JSON, multipart, and helper modules.

This keeps the install helpers out of the public crate API and removes
stringly typed source dispatch while preserving existing install
behaviour.
