+++ src/skills/mod.rs
@@ -19,7 +19,7 @@ pub mod bundle;
 pub mod catalog;
 pub mod gating;
 /// Shared source-field normalisation helpers for skill install adapters.
-pub mod install_source;
+pub(crate) mod install_source;
 pub mod parser;
 pub mod registry;
 pub mod selector;
